Add files via upload
To further improve the performance of the online portfolio considering transaction costs, we propose a transaction cost regularization (TCR) model. This model minimizes the negative expected return and meanwhile incorporates the L1 norm of the difference between two consecutive allocations as a regularization term to control the transaction cost. We first apply the linearized augmented Lagrangian method to solve the proposed model and derive a convergent iterative algorithm whose update in each iteration has a closed-form, which enable the computational efficiency of the algorithm. To further study the problem, the alternative direction method of multipliers is applied to solve the portfolio model and its convergence is theoretically guaranteed as well as guaranteeing that the portfolio variables for each iteration are constrained in simplex. Numerical experiments on benchmark datasets illustrate that our proposed algorithms are efficient and can achieve higher cumulative wealth than the compared state-ofthe-art algorithms in most cases.